requirements.txt
anyio==4.9.0
APScheduler==3.11.0
aria2p==0.12.1
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.3
packaging==25.0
platformdirs==4.3.8
python-dotenv==1.1.0
python-telegram-bot==22.1
PyYAML==6.0.2
qbittorrent-api==2025.5.0
requests==2.32.3
sniffio==1.3.1
transmission-rpc==7.0.11
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
